Confirm release belongs to HR before upgrading
Before this change, when a user would create multiple HelmReleases
with the same release name configured, the operator would attempt
to upgrade the release indefinitely.

To overcome this issue, the operator now injects the resource ID
of the HelmRelease into the description of the release and confirms
this value equals the resource ID of the HelmRelease it is going to
run an upgrade for. If the value diverges, the status of the
HelmRelease is updated to make it visible to the user and a message
is logged.

For backwards compatibility, and to be able to migrate existing
releases to a HelmRelease, we see empty descriptions as a positive.